🚚 move: App 디렉터리 이동
