+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './example/App';
+import App from './components/App';
 import reportWebVitals from './util/reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
